Move CKRepresentable conformance to separate files
Relocated EventObject and SessionObject CKRepresentable implementations from CKPersistable.swift to dedicated extension files for better code organization and separation of concerns.
