+++ Sources/Scout/Core/Sync/CKPersistable.swift
@@ -16,44 +16,3 @@ protocol CKInitializable {
 protocol CKRepresentable {
     var toRecord: CKRecord { get }
 }
-
-extension EventObject: CKRepresentable {
-    var toRecord: CKRecord {
-        let record = CKRecord(recordType: "Event")
-
-        record["name"] = name
-        record["level"] = level
-        record["params"] = params
-        record["param_count"] = paramCount
-
-        record["date"] = date
-        record.setValuesForKeys(dateFields)
-
-        record["uuid"] = eventID?.uuidString
-        record["session_id"] = sessionID?.uuidString
-        record["launch_id"] = launchID?.uuidString
-        record["user_id"] = userID?.uuidString
-
-        record["version"] = 1
-
-        return record
-    }
-}
-
-extension SessionObject: CKRepresentable {
-    var toRecord: CKRecord {
-        let record = CKRecord(recordType: "Session")
-
-        record["start_date"] = date
-        record["end_date"] = endDate
-        record.setValuesForKeys(dateFields)
-
-        record["session_id"] = sessionID?.uuidString
-        record["launch_id"] = launchID?.uuidString
-        record["user_id"] = userID?.uuidString
-
-        record["version"] = 1
-
-        return record
-    }
-}
